modify ASRFactoryDataProvider.c to C++
