+++ src/platform/ASR/BUILD.gn
@@ -26,7 +26,7 @@ static_library("ASR") {
     "../SingletonConfigurationManager.cpp",
     "ASRConfig.cpp",
     "ASRConfig.h",
-    "ASRFactoryDataParser.c",
+    "ASRFactoryDataParser.cpp",
     "ASRFactoryDataParser.h",
     "ASRFactoryDataProvider.cpp",
     "ASRUtils.cpp",
